non working version for species structure
Separate species into autotroph and heterotroph sub-classes. Changes the database accordingly. Now still need to add interface to add the new version of a species, as i also added different fields and meta fields that decide the fitness of the species based on its environment(for now simply the other populations and their species present in the same region as themselves), and then cahnge the way fitness is decided( division of prey species into different pools)